delete room-photo.py (doesn't using) add models/host_booking.py (model of self booking room by host) update models/__init__.py (import update)
