+++ server/models/__init__.py
@@ -2,3 +2,4 @@ from models.user import UserModel
 from models.store import StoreModel
 from models.room import RoomModel
 from models.room_photo import RoomPhotoModel
+from models.host_booking import HostSelfbookingModel
